requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic-settings>=2.0.0
pymupdf>=1.24.0
python-docx>=1.1.0
google-cloud-aiplatform>=1.50.0
google-genai>=1.0.0
google-cloud-firestore>=2.16.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
streamlit>=1.35.0
requests>=2.31.0
